Moved game related classes to new package in JungleUtil
