+++ server/src/main/java/edu/colostate/cs/cs414/chesshireCoders/jungleServer/session/LoginVerifier.java
@@ -2,8 +2,8 @@ package edu.colostate.cs.cs414.chesshireCoders.jungleServer.session;
 
 import edu.colostate.cs.cs414.chesshireCoders.jungleServer.dataAccessObjects.LoginDAO;
 import edu.colostate.cs.cs414.chesshireCoders.jungleServer.dataAccessObjects.UserDAO;
-import edu.colostate.cs.cs414.chesshireCoders.jungleServer.dataObjects.Login;
-import edu.colostate.cs.cs414.chesshireCoders.jungleServer.dataObjects.User;
+import edu.colostate.cs.cs414.chesshireCoders.jungleUtil.game.Login;
+import edu.colostate.cs.cs414.chesshireCoders.jungleUtil.game.User;
 
 import java.sql.SQLException;
