+++ util/src/main/java/edu/colostate/cs/cs414/chesshireCoders/jungleUtil/game/Player.java
@@ -1,4 +1,4 @@
-package edu.colostate.cs.cs414.chesshireCoders.jungleServer.dataObjects;
+package edu.colostate.cs.cs414.chesshireCoders.jungleUtil.game;
 
 import edu.colostate.cs.cs414.chesshireCoders.jungleUtil.types.GameOutcomeType;
 import edu.colostate.cs.cs414.chesshireCoders.jungleUtil.types.PlayerColor;
